fix shareTo feature
it was broken since ConversationsList was migrated to Activity
