+++ app/src/main/AndroidManifest.xml
@@ -54,8 +54,8 @@
     <uses-permission android:name="android.permission.CHANGE_WIFI_MULTICAST_STATE" />
     <uses-permission android:name="android.permission.MODIFY_AUDIO_SETTINGS" />
     <uses-permission android:name="android.permission.VIBRATE" />
-    <uses-permission android:name="android.permission.USE_FULL_SCREEN_INTENT"/>
-    <uses-permission android:name="android.permission.FOREGROUND_SERVICE"/>
+    <uses-permission android:name="android.permission.USE_FULL_SCREEN_INTENT" />
+    <uses-permission android:name="android.permission.FOREGROUND_SERVICE" />
     <uses-permission android:name="android.permission.READ_CONTACTS" />
     <uses-permission android:name="android.permission.WRITE_CONTACTS" />
     <uses-permission android:name="android.permission.READ_PROFILE" />
@@ -71,7 +71,8 @@
         android:name="android.permission.WRITE_EXTERNAL_STORAGE"
         android:maxSdkVersion="29"
         tools:ignore="ScopedStorage" />
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"
+    <uses-permission
+        android:name="android.permission.READ_EXTERNAL_STORAGE"
         android:maxSdkVersion="32" />
     <uses-permission android:name="android.permission.READ_MEDIA_IMAGES" />
     <uses-permission android:name="android.permission.READ_MEDIA_VIDEO" />
@@ -89,7 +90,7 @@
     <permission
         android:name="${applicationId}.${broadcastPermission}"
         android:protectionLevel="signature" />
-    <uses-permission android:name="${applicationId}.${broadcastPermission}"/>
+    <uses-permission android:name="${applicationId}.${broadcastPermission}" />
 
     <application
         android:name=".application.NextcloudTalkApplication"
@@ -99,9 +100,9 @@
         android:label="@string/nc_app_name"
         android:largeHeap="true"
         android:networkSecurityConfig="@xml/network_security_config"
+        android:requestLegacyExternalStorage="true"
         android:supportsRtl="true"
         android:theme="@style/AppTheme.Launcher"
-        android:requestLegacyExternalStorage="true"
         tools:ignore="UnusedAttribute"
         tools:replace="label, icon, theme, name, allowBackup">
 
@@ -119,20 +120,9 @@
             </intent-filter>
 
             <intent-filter>
-                <action android:name="android.intent.action.SEND" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <data android:mimeType="*/*" />
-            </intent-filter>
-
-            <intent-filter>
-                <action android:name="android.intent.action.SEND_MULTIPLE" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <data android:mimeType="*/*" />
-            </intent-filter>
-
-            <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
                 <category android:name="android.intent.category.DEFAULT" />
+
                 <data android:mimeType="vnd.android.cursor.item/vnd.com.nextcloud.talk2.chat" />
                 <data android:scheme="content" />
                 <data android:scheme="file" />
@@ -141,41 +131,38 @@
 
         <activity
             android:name=".activities.CallActivity"
-            android:theme="@style/AppTheme.CallLauncher"
-            android:supportsPictureInPicture="true"
             android:configChanges="screenSize|smallestScreenSize|screenLayout|orientation"
-            android:launchMode="singleTask"
-            android:taskAffinity=".call"
             android:excludeFromRecents="true"
-            android:showOnLockScreen="true"/>
+            android:launchMode="singleTask"
+            android:showOnLockScreen="true"
+            android:supportsPictureInPicture="true"
+            android:taskAffinity=".call"
+            android:theme="@style/AppTheme.CallLauncher" />
 
         <activity
             android:name=".activities.CallNotificationActivity"
-            android:theme="@style/AppTheme.CallLauncher"
-            android:supportsPictureInPicture="true"
             android:configChanges="screenSize|smallestScreenSize|screenLayout|orientation"
-            android:launchMode="singleTask"
-            android:taskAffinity=".call"
             android:excludeFromRecents="true"
-            android:showOnLockScreen="true" />
+            android:launchMode="singleTask"
+            android:showOnLockScreen="true"
+            android:supportsPictureInPicture="true"
+            android:taskAffinity=".call"
+            android:theme="@style/AppTheme.CallLauncher" />
 
         <activity
             android:name=".activities.FullScreenImageActivity"
-            android:theme="@style/FullScreenImageTheme"
-            android:configChanges="orientation|keyboardHidden|screenSize">
-        </activity>
+            android:configChanges="orientation|keyboardHidden|screenSize"
+            android:theme="@style/FullScreenImageTheme"></activity>
 
         <activity
             android:name=".activities.FullScreenMediaActivity"
-            android:theme="@style/FullScreenMediaTheme"
-            android:configChanges="orientation|keyboardHidden|screenSize">
-        </activity>
+            android:configChanges="orientation|keyboardHidden|screenSize"
+            android:theme="@style/FullScreenMediaTheme"></activity>
 
         <activity
             android:name=".activities.FullScreenTextViewerActivity"
-            android:theme="@style/FullScreenTextTheme"
-            android:configChanges="orientation|keyboardHidden|screenSize">
-        </activity>
+            android:configChanges="orientation|keyboardHidden|screenSize"
+            android:theme="@style/FullScreenTextTheme"></activity>
 
         <activity
             android:name=".activities.TakePhotoActivity"
@@ -224,16 +211,31 @@
 
         <activity
             android:name=".chat.ChatActivity"
-            android:theme="@style/AppTheme"
             android:launchMode="singleInstance"
             android:noHistory="true"
-            android:screenOrientation="portrait" />
+            android:screenOrientation="portrait"
+            android:theme="@style/AppTheme" />
 
         <activity
             android:name=".conversationlist.ConversationsListActivity"
-            android:theme="@style/AppTheme" />
+            android:theme="@style/AppTheme"
+            android:exported="true">
 
-        <receiver android:name=".receivers.PackageReplacedReceiver"
+            <intent-filter>
+                <action android:name="android.intent.action.SEND" />
+                <category android:name="android.intent.category.DEFAULT" />
+                <data android:mimeType="*/*" />
+            </intent-filter>
+
+            <intent-filter>
+                <action android:name="android.intent.action.SEND_MULTIPLE" />
+                <category android:name="android.intent.category.DEFAULT" />
+                <data android:mimeType="*/*" />
+            </intent-filter>
+        </activity>
+
+        <receiver
+            android:name=".receivers.PackageReplacedReceiver"
             android:exported="false">
             <intent-filter>
                 <action android:name="android.intent.action.MY_PACKAGE_REPLACED" />
@@ -260,7 +262,8 @@
                 android:resource="@xml/contacts" />
         </service>
 
-        <service android:name=".utils.AuthenticatorService"
+        <service
+            android:name=".utils.AuthenticatorService"
             android:exported="false">
             <intent-filter>
                 <action android:name="android.accounts.AccountAuthenticator" />
