+++ app/src/main/java/com/nextcloud/talk/ui/dialog/ChooseAccountShareToDialogFragment.kt
@@ -27,6 +27,7 @@ import android.annotation.SuppressLint
 import android.app.Dialog
 import android.net.Uri
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -34,7 +35,6 @@ import androidx.fragment.app.DialogFragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import autodagger.AutoInjector
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
-import com.nextcloud.talk.activities.MainActivity
 import com.nextcloud.talk.adapters.items.AdvancedUserItem
 import com.nextcloud.talk.application.NextcloudTalkApplication
 import com.nextcloud.talk.application.NextcloudTalkApplication.Companion.sharedApplication
@@ -46,6 +46,8 @@ import com.nextcloud.talk.ui.theme.ViewThemeUtils
 import com.nextcloud.talk.users.UserManager
 import eu.davidea.flexibleadapter.FlexibleAdapter
 import eu.davidea.flexibleadapter.common.SmoothScrollLinearLayoutManager
+import io.reactivex.SingleObserver
+import io.reactivex.disposables.Disposable
 import java.net.CookieManager
 import javax.inject.Inject
 
@@ -160,7 +162,37 @@ class ChooseAccountShareToDialogFragment : DialogFragment() {
             val user = userItems[position].user
             if (userManager!!.setUserAsActive(user).blockingGet()) {
                 cookieManager!!.cookieStore.removeAll()
-                activity?.runOnUiThread { (activity as MainActivity?)!!.resetConversationsList() }
+                // activity?.runOnUiThread { (activity as MainActivity?)!!.resetConversationsList() }
+
+                userManager!!.users.subscribe(object : SingleObserver<List<User>> {
+                    override fun onSubscribe(d: Disposable) {
+                        // unused atm
+                    }
+
+                    override fun onSuccess(users: List<User>) {
+                        if (users.isNotEmpty()) {
+                            // runOnUiThread {
+                            //     setDefaultRootController()
+                            // }
+
+                            // val intent = Intent(activity, ConversationsListActivity::class.java)
+                            //
+                            // // val intent = Intent(context, ConversationsListActivity::class.java)
+                            // intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+                            // activity?.intent?.extras?.let { intent.putExtras(it) }
+                            // startActivity(intent)
+
+                            // startActivity(activity?.intent)
+
+                            activity?.recreate()
+                        }
+                    }
+
+                    override fun onError(e: Throwable) {
+                        Log.e(TAG, "Error loading existing users", e)
+                    }
+                })
+
                 dismiss()
             }
         }
